implement DEL position with stock calculatio. closes issue#5

+++ models/positions.py
@@ -16,6 +16,7 @@ class PositionsModel(db.Model):  # extend db.Model for SQLAlechemy
     JSON_DATE_STR = 'date'
     JSON_QUANTITY_STR = 'quantity'
     JSON_UNIT_COST_STR = 'unit_cost'
+    JSON_POSITION_ID_STR = 'position_id'
 
     __tablename__ = 'positions'
 
@@ -23,7 +24,7 @@ class PositionsModel(db.Model):  # extend db.Model for SQLAlechemy
     quantity = db.Column(db.Integer)
     position_date = db.Column(db.Date)
     unit_cost = db.Column(db.Float)
-    #unit_cost = db.Column(db.Float(precision=const.PRICE_PRECISION))
+    # unit_cost = db.Column(db.Float(precision=const.PRICE_PRECISION))
     calc_flag = db.Column(db.Boolean)  # indicator for calc
     stock_id = db.Column(db.Integer, db.ForeignKey('stock.id'), nullable=False)  # foreign key to stock table
 
@@ -70,6 +71,22 @@ class PositionsModel(db.Model):  # extend db.Model for SQLAlechemy
         return parser.parse_args(strict=True)  # only the one argument can be in the request
 
     @classmethod
+    def parse_request_json_position_id(cls):
+        """
+        parse position id from request
+        {"position_id": <position_id>
+        """
+        parser = reqparse.RequestParser()
+        parser.add_argument(
+            name=PositionsModel.JSON_POSITION_ID_STR,
+            type=int,
+            required=True,
+            trim=True,
+            help='position id field is missing or invalid')
+        current_app.logger.debug('func: parse_request_json_position_id, parse={}'.format(dict(parser.parse_args())))
+        return parser.parse_args(strict=True)  # only the one argument can be in the request
+
+    @classmethod
     def find_by_symbol(cls, symbol):
         """
         find record in DB according to symbol
@@ -80,7 +97,16 @@ class PositionsModel(db.Model):  # extend db.Model for SQLAlechemy
             stock_id = stock.id
         else:
             return None
-        return cls.query.filter_by(stock_id=stock_id).order_by(PositionsModel.position_date).all()  # SQLAlchemy -> SELECT * FROM stock WHERE symbol=symbol
+        return cls.query.filter_by(stock_id=stock_id).order_by(
+            PositionsModel.position_date).all()  # SQLAlchemy -> SELECT * FROM position WHERE stock_id=stock_id
+
+    @classmethod
+    def find_by_position_id(cls, position_id):
+        """
+        find record in DB according to symbol
+        if found, return object with stock details, otherwise None
+        """
+        return cls.query.get(position_id)  # SQLAlchemy -> SELECT * FROM position WHERE id=position_id
 
     def json(self) -> dict:
         """
@@ -99,25 +125,34 @@ class PositionsModel(db.Model):  # extend db.Model for SQLAlechemy
         """
         insert position details for a stock
         """
-        self.stock_id = StockModel.find_by_symbol(self.symbol).id
+        stock = StockModel.find_by_symbol(self.symbol)
+        self.stock_id = stock.id
         current_app.logger.debug('func: save_details, self={}'.format(self.json()))
         db.session.add(self)
-        db.session.commit()
-        current_app.logger.debug('func: save_details, after commit')
-        # adding for calc unit_cost and quantity. edit properly TODO
-        stock = StockModel.find_by_symbol(self.symbol)
+        # find the stock and update its unit_cost and quantity
         current_app.logger.debug('func: save_details, stock={}'.format(stock))
+        # calc by adding unit_cost and quantity to existing
         stock.calc_unit_cost_and_quantity(self.unit_cost, self.quantity)
-        self.calc_flag = True
+        self.calc_flag = True  # set the flag to show position got calculated
         db.session.commit()
 
-
-    def del_position(self):
+    def del_position(self, symbol) -> bool:
         """
         delete stock from DB
+        :return True for success, False for failure
         """
-        db.session.delete(self)
-        db.session.commit()
+        # find the stock and calculate the updated unit_cost and quantity
+        stock = StockModel.find_by_symbol(symbol)
+        if stock.id == self.stock_id:
+            db.session.delete(self)
+            current_app.logger.debug('func: del_position, stock={}'.format(stock.detailed_json()))
+            stock.calc_unit_cost_and_quantity(self.unit_cost, -self.quantity)
+            db.session.commit()
+            return True
+        else:
+            # not match between position stock id and the stock symbol
+            return False
+
 
     def get_stock_id(self):
         """
